Obtain type descriptor via `TypeUrl`
